Add global declarations
This allows for spell checking on all arguments.

This breaks basic recursion as seen in the fib test however. This
will be addressed when uptable values are added, as globals and
locals will be inherited from their environment.

The tests have been temporarily updated to work under the constraints
of the current scoping system. This will be fixed once a better c interface
to add global variables to the scope is in place.